+++ src/SecuredApp.tsx
@@ -19,7 +19,9 @@ const SecuredAppComponent = ({ AppComponent }: SecuredAppProps) => {
   }, [getAccessTokenSilently]);
 
   return (
-    <>
+    <div className="ctw-space-y-5">
+      {authToken && <AppComponent accessToken={authToken} />}
+
       <button
         type="button"
         onClick={() => logout({ returnTo })}
@@ -27,9 +29,7 @@ const SecuredAppComponent = ({ AppComponent }: SecuredAppProps) => {
       >
         Log Out
       </button>
-
-      {authToken && <AppComponent accessToken={authToken} />}
-    </>
+    </div>
   );
 };