+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # ctw-component-library
 
+## 1.61.1
+
+### Patch Changes
+
+- c44b9d55: Fixed PatientRecordSearch results heading and vertical spacing styles.
+
+## 1.61.0
+
+### Minor Changes
+
+- 965ec255: Added AISearch poc component
+
+## 1.60.5
+
+### Patch Changes
+
+- 85b80dca: Export fqs client.
+
 ## 1.60.4
 
 ### Patch Changes